+++ src/ingest_validation_tools/enums.py
@@ -8,55 +8,53 @@ from typing import Dict, List
 >>> for name in local_names:
 ...     response = requests.post(
 ...         'https://search.api.hubmapconsortium.org/assayname',
-...         json={'name':name}, headers={'Content-Type': 'application/json'}).json()
+...         json={'name': name},
+...         headers={'Content-Type': 'application/json'}
+...     ).json()
 ...     if 'error' in response:
 ...         remote_mismatch.append(name)
->>> print('\\n'.join(remote_mismatch))
-<BLANKLINE>
+>>> print(remote_mismatch)
+[]
 '''
-# The list above should be empty: That means that all the assays
-# listed below are recognized by the assay service.
-# TODO: Joel will progressively uncomment names below, and when that is done
-#       _validate_level_1_enum() can be renabled in schema_loader.py.
-#       https://github.com/hubmapconsortium/ingest-validation-tools/issues/1023
+# The list above should be empty: There should be no mismatches;
+# ie: All the assays listed below are recognized by the assay service.
+#
+# The assay_type list is *all* the values which have ever been used,
+# including schemas which are currently deprecated.
+# Each schema lists the particular values which are valid for it.
+# This redundant list catches mistakes like a typo being introduced
+# when an existing schema is copied, pasted, and tweaked to make a new version.
 
 shared_enums: Dict[str, List[str]] = {
-    'assay_category': [
-        'imaging',
-        'mass_spectrometry',
-        'mass_spectrometry_imaging',
-        'sequence'
-    ],
     'assay_type': [
         '3D Imaging Mass Cytometry',
-        'scRNA-Seq(10xGenomics)',
         'AF',
-        # 'bulk RNA',
+        'bulk RNA',
         'bulkATACseq',
         'Cell DIVE',
-        # 'CE-MS',
+        'CE-MS',
         'CODEX',
-        # 'DESI',
-        # 'GC-MS',
+        'CODEX2',  # TODO: Temporary; will be removed.
+        # https://github.com/hubmapconsortium/ingest-validation-tools/issues/1107
+        'DESI',
+        'GC-MS',
         'Imaging Mass Cytometry',
-        # 'LC-MS (metabolomics)',
+        'LC-MS (metabolomics)',
         'LC-MS/MS (label-free proteomics)',
         'Light Sheet',
         'MxIF',
         'MALDI-IMS',
-        # 'MS (shotgun lipidomics)',
+        'MS (shotgun lipidomics)',
         'Multiplex Ion Beam Imaging',
-        # 'NanoDESI',
-        # 'NanoPOTS',
+        'NanoDESI',
+        'NanoPOTS',
         'PAS microscopy',
         'sciATACseq',
         'sciRNAseq',
         'seqFISH',
-        # 'SIMS-IMS',
+        'SIMS-IMS',
         'SNARE-seq2',
         'snATACseq',
-        # 'SPLiT-Seq',
-        # 'TMT (proteomics)',
         'WGS',
         'SNARE2-RNAseq',
         'snRNAseq',  # equivalent to snRNAseq-10xGenomics-v3
@@ -65,14 +63,19 @@ shared_enums: Dict[str, List[str]] = {
         'scRNAseq-10xGenomics-v3',
         'snRNAseq-10xGenomics-v2',
         'snRNAseq-10xGenomics-v3',
-        # 'scRNAseq',
+        'scRNAseq',
         'Slide-seq',
-        # 'MS Bottom-Up',
-        # 'MS Top-Down',
-        # 'LC-MS Top-Down',
-        # 'LC-MS',
-        # 'LC-MS Bottom-Up',
-        # 'MS'
+        'MS Bottom-Up',
+        'MS Top-Down',
+        'LC-MS Top-Down',
+        'LC-MS',
+        'LC-MS Bottom-Up',
+        'MS',
+        'Body CT',
+        'Micro CT',
+        'OCT',
+        'MRI',
+        'Ultrasound',
     ],
     'analyte_class': [
         'DNA',
@@ -85,5 +88,12 @@ shared_enums: Dict[str, List[str]] = {
         'glycans',
         'peptides',
         'phosphopeptides'
-    ]
+    ],
+    'assay_category': [
+        'clinical_imaging',
+        'imaging',
+        'mass_spectrometry',
+        'mass_spectrometry_imaging',
+        'sequence'
+    ],
 }